- Removed do_reasoning argument from tagsistant_querytree_new(), being     obsoleted by the '@@' operator.
  - Solved two bugs in deduplication which prevented real unlink() of the
    file from disk and leaved the old inode stored in AND_SET cache.


M    src/buildnumber.h
M    src/path_resolution.c
M    src/tagsistant.c
M    src/fuse_operations/readlink.c
M    src/fuse_operations/chmod.c
M    src/fuse_operations/write.c
M    src/fuse_operations/link.c
M    src/fuse_operations/unlink.c
M    src/fuse_operations/open.c
M    src/fuse_operations/mkdir.c
M    src/fuse_operations/rename.c
M    src/fuse_operations/mknod.c
M    src/fuse_operations/getattr.c
M    src/fuse_operations/readdir.c
M    src/fuse_operations/read.c
M    src/fuse_operations/rmdir.c
M    src/fuse_operations/chown.c
M    src/fuse_operations/release.c
M    src/fuse_operations/flush.c
M    src/fuse_operations/utime.c
M    src/fuse_operations/truncate.c
M    src/fuse_operations/symlink.c
M    src/path_resolution.h
M    src/utils.c
M    ChangeLog

+++ src/tagsistant.c
@@ -323,7 +323,7 @@ int main(int argc, char *argv[])
 	if (tagsistant.debug) {
 		char *dbg_ptr = tagsistant.debug;
 		while (*dbg_ptr) {
-			tagsistant.dbg[*dbg_ptr] = 1;
+			tagsistant.dbg[(int) *dbg_ptr] = 1;
 			dbg_ptr++;
 		}
 	}
